scripts/requirements.txt
# Image processing for Boing assets (background modifier + hero element extraction)
Pillow>=10.0.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
rembg>=2.0.50
